fix bugs.

+++ install.sh
@@ -19,16 +19,18 @@ SOURCE_TEMPLATE_PROJECT="$SOURCE_TEMPLATE/$TEMPLATE_PROJECT/$BEEFRAMEWORK"
 # # 如果已有，先删除上一个版本的
 if [ -d "$TEMPLATE_FILE_PATH/$BEEFRAMEWORK" ]; then
   sudo rm -rf "$TEMPLATE_FILE_PATH/$BEEFRAMEWORK"
-  echo "$TEMPLATE_FILE_PATH/$BEEFRAMEWORK"
+  echo "remove $TEMPLATE_FILE_PATH/$BEEFRAMEWORK"
 fi
 
 if [ -d "$TEMPLATE_PROJECT_PATH/$BEEFRAMEWORK" ]; then
   sudo rm -rf "$TEMPLATE_PROJECT_PATH/$BEEFRAMEWORK"
-  echo "$TEMPLATE_PROJECT_PATH/$BEEFRAMEWORK"
+  echo "remove $TEMPLATE_PROJECT_PATH/$BEEFRAMEWORK"
 fi
 
 # # 从网络下载
-status=$(sudo curl -L "$PACKAGE_URL" -o "$PACKAGE")
+if [[ ! -f "$PACKAGE" ]]; then
+  status=$(sudo curl -L "$PACKAGE_URL" -o "$PACKAGE")
+fi
 
 # 解压
 echo "unzip package ..."
@@ -36,11 +38,25 @@ sudo tar xzf "$PACKAGE"
 
 # 复制
 echo "copy templates ..."
+
+if [[ ! -d "$TEMPLATE_FILE_PATH" ]]; then
+  sudo mkdir -p "$TEMPLATE_FILE_PATH"
+  echo "$TEMPLATE_FILE_PATH"
+fi
+
 sudo cp -rf "$SOURCE_TEMPLATE_FILE" "$TEMPLATE_FILE_PATH"
+
+if [[ ! -d "$TEMPLATE_PROJECT_PATH" ]]; then
+  sudo mkdir -p "$TEMPLATE_PROJECT_PATH"
+  echo "$TEMPLATE_PROJECT_PATH"
+fi
+
 sudo cp -rf "$SOURCE_TEMPLATE_PROJECT" "$TEMPLATE_PROJECT_PATH"
 
 # cleanup
 # sudo rm -rf "$PACKAGE"
 sudo rm -rf "Bee-Xcode-Template-master"
 
-echo "Bee Xcode Template is installed."
+echo "
+# Bee Xcode Template is installed to $TEMPLATE_PATH. 
+# "
